+++ Tests/DependencyInjection/FirebaseExtensionTest.php
@@ -4,9 +4,14 @@ declare(strict_types=1);
 
 namespace Kreait\Firebase\Symfony\Bundle\Tests\DependencyInjection;
 
+use Kreait\Firebase;
 use Kreait\Firebase\Symfony\Bundle\DependencyInjection\FirebaseExtension;
 use PHPUnit\Framework\TestCase;
+use Symfony\Component\Config\Definition\Exception\InvalidConfigurationException;
+use Symfony\Component\DependencyInjection\Alias;
+use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Definition;
 
 class FirebaseExtensionTest extends TestCase
 {
@@ -45,7 +50,7 @@ class FirebaseExtensionTest extends TestCase
             'projects' => [
                 'foo' => [
                     'alias' => 'bar',
-                    'public' => false
+                    'public' => false,
                 ],
             ],
         ]);
@@ -86,9 +91,91 @@ class FirebaseExtensionTest extends TestCase
         $this->assertTrue($container->hasDefinition($this->extension->getAlias().'.foo'));
     }
 
-    protected function createContainer(array $config = [])
+    /**
+     * @test
+     */
+    public function it_accepts_only_one_default_project()
+    {
+        $this->expectException(InvalidConfigurationException::class);
+
+        $this->createContainer([
+            'projects' => [
+                'foo' => [
+                    'default' => true,
+                ],
+                'bar' => [
+                    'default' => true,
+                ],
+            ],
+        ]);
+    }
+
+    /**
+     * @test
+     */
+    public function it_aliases_the_firebase_class_to_the_default_project()
+    {
+        $container = $this->createContainer([
+            'projects' => [
+                'foo' => [],
+                'bar' => [
+                    'default' => true,
+                ],
+            ],
+        ], $makeServicesPublic = true);
+
+        $this->assertTrue($container->hasAlias(Firebase::class));
+    }
+
+    /**
+     * @test
+     */
+    public function it_aliases_the_firebase_class_to_the_only_project()
+    {
+        $container = $this->createContainer([
+            'projects' => [
+                'foo' => [],
+            ],
+        ], $makeServicesPublic = true);
+
+        $this->assertTrue($container->hasAlias(Firebase::class));
+    }
+
+    /**
+     * @test
+     */
+    public function it_has_no_default_project_if_none_could_be_determined()
+    {
+        $container = $this->createContainer([
+            'projects' => [
+                'foo' => [],
+                'bar' => [],
+            ],
+        ], $makeServicesPublic = true);
+
+        $this->assertFalse($container->hasAlias(Firebase::class));
+    }
+
+    protected function createContainer(array $config = [], $makeServicesPublic = false)
     {
         $container = new ContainerBuilder();
+
+        // Make all services public just for testing
+        if ($makeServicesPublic) {
+            $container->addCompilerPass(new class implements CompilerPassInterface
+            {
+                public function process(ContainerBuilder $container)
+                {
+                    array_map(function (Definition $definition) {
+                        $definition->setPublic(true);
+                    }, $container->getDefinitions());
+
+                    array_map(function (Alias $alias) {
+                        $alias->setPublic(true);
+                    }, $container->getAliases());
+                }
+            });
+        }
 
         $this->extension->load([$this->extension->getAlias() => $config], $container);